+++ lib/src/fx_suit/fx_components.dart
@@ -1,48 +1,54 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_fx/src/fx_suit/fx_extensions.dart';
+import 'package:flutter/widgets.dart';
 
 class FxCustomTextField extends StatelessWidget {
-  final bool hasShadow;
+  final double shadowBlurRadius;
+  final Offset shadowOffset;
   final String? label;
   final bool? autoFocus;
-  final double? maxWidth;
-  final Color? fillColor;
+  final Color backgroundColor;
   final bool? obscureText;
   final bool? enableInput;
+  final int? maxLines;
   final String? errorText;
   final Widget? suffixIcon;
-  final EdgeInsets? margin;
   final Widget? prefixIcon;
   final FocusNode? focusNode;
   final String? placeHolder;
   final TextStyle? textStyle;
-  final TextStyle? hintStyle;
   final TextStyle? errorStyle;
   final EdgeInsets? inputPadding;
   final TextInputType? inputType;
-  final Alignment? alignment;
   final InputBorder? inputBorder;
-  final BoxConstraints? constraints;
+  final double? maxWidth;
+  final double? maxHeight;
+  final Color? cursorColor;
+  final bool autoUnfocus;
+  final EdgeInsetsGeometry margin;
   final TextEditingController? inputController;
   final Function(String)? onChanged;
   final Function(String)? onSubmitted;
 
+  final TextStyle defaultTextStyle = const TextStyle(
+      height: 1, fontSize: 15, color: Color.fromARGB(255, 204, 204, 204));
+
   const FxCustomTextField({
     super.key,
     required this.inputController,
-    this.constraints,
-    this.hasShadow = false,
+    this.cursorColor,
+    this.maxHeight,
+    this.maxWidth,
+    this.autoUnfocus = false,
+    this.maxLines = 1,
+    this.shadowBlurRadius = 10.0,
+    this.shadowOffset = const Offset(0, 5),
     this.placeHolder,
     this.focusNode,
-    this.alignment,
     this.inputType,
     this.onChanged,
     this.onSubmitted,
-    this.margin,
-    this.maxWidth,
     this.inputPadding,
     this.textStyle,
-    this.hintStyle,
     this.label,
     this.errorStyle,
     this.enableInput,
@@ -52,92 +58,74 @@ class FxCustomTextField extends StatelessWidget {
     this.obscureText,
     this.suffixIcon,
     this.prefixIcon,
-    this.fillColor,
+    this.margin = const EdgeInsets.symmetric(vertical: 5.0),
+    this.backgroundColor = const Color.fromARGB(255, 96, 96, 96),
   });
 
-  final Color white = const Color(0xFFFFFFFF);
+  final Color warning = const Color.fromARGB(255, 124, 42, 42);
+  final Color black = const Color.fromARGB(255, 0, 0, 0);
+  final Color white = const Color.fromARGB(255, 255, 255, 255);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      alignment: alignment ?? Alignment.center,
-      child: Container(
-          margin: margin ??
-              EdgeInsets.symmetric(vertical: 7.0.scaleSize, horizontal: 0),
-          padding: const EdgeInsets.all(0.0),
-          decoration: BoxDecoration(
-            boxShadow: hasShadow
-                ? [
-                    const BoxShadow(
-                        color: Color(0x19000000),
-                        offset: Offset(0, 5),
-                        blurRadius: 6,
-                        spreadRadius: 0.2,
-                        blurStyle: BlurStyle.normal)
-                  ]
-                : null,
+    return Padding(
+      padding: margin,
+      child: TextField(
+          textInputAction: TextInputAction.done,
+          onTapOutside: (event) {
+            if (autoUnfocus) {
+              focusNode?.unfocus();
+            }
+          },
+          decoration: InputDecoration(
+            errorText: errorText,
+            label: (label != null) ? Text(label!) : null,
+            errorStyle: errorStyle ?? defaultTextStyle.copyWith(color: warning),
+            fillColor: backgroundColor,
+            filled: true,
+            focusedBorder: inputBorder ??
+                OutlineInputBorder(
+                  borderSide: BorderSide(color: backgroundColor),
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+            enabledBorder: inputBorder ??
+                OutlineInputBorder(
+                  borderSide: BorderSide(color: backgroundColor),
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+            border: inputBorder ??
+                OutlineInputBorder(
+                  borderSide: BorderSide(color: backgroundColor),
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+            isDense: true,
+            contentPadding: inputPadding ??
+                const EdgeInsets.symmetric(horizontal: 15.0, vertical: 12.0),
+            constraints: BoxConstraints(
+                maxWidth: maxWidth ?? double.maxFinite,
+                minHeight: 40.0,
+                maxHeight: maxHeight ?? 90.0),
+            hintText: placeHolder,
+            prefixIcon: prefixIcon,
+            hintStyle: textStyle ?? defaultTextStyle,
+            labelStyle: textStyle ?? defaultTextStyle,
+            suffixIconConstraints: const BoxConstraints(maxWidth: 60.0),
+            iconColor: white,
+            suffixIcon: suffixIcon,
           ),
-          child: TextField(
-              textAlign: TextAlign.left,
-              textAlignVertical: TextAlignVertical.center,
-              decoration: InputDecoration(
-                errorText: errorText,
-                alignLabelWithHint: true,
-                label: (label != null) ? Text(label!) : null,
-                errorStyle: textStyle,
-                fillColor: fillColor ?? white,
-                filled: true,
-                isDense: true,
-                isCollapsed: true,
-                focusedBorder: inputBorder ??
-                    OutlineInputBorder(
-                      borderSide: BorderSide(color: fillColor ?? white),
-                      borderRadius: BorderRadius.circular(6.0),
-                    ),
-                enabledBorder: inputBorder ??
-                    OutlineInputBorder(
-                      borderSide: BorderSide(color: fillColor ?? white),
-                      borderRadius: BorderRadius.circular(6.0),
-                    ),
-                border: inputBorder ??
-                    OutlineInputBorder(
-                      borderSide: BorderSide(color: fillColor ?? white),
-                      borderRadius: BorderRadius.circular(6.0),
-                    ),
-                hintText: placeHolder,
-                prefixIcon: prefixIcon,
-                contentPadding: inputPadding ??
-                    EdgeInsets.symmetric(
-                        horizontal: 15.0.scaleSize, vertical: 12.0.scaleSize),
-                constraints: constraints ??
-                    BoxConstraints(
-                        maxWidth: 260.0.scaleSize,
-                        maxHeight: 40.0.scaleSize,
-                        minHeight: 40.0.scaleSize),
-                prefixIconConstraints: BoxConstraints(
-                    minWidth: 10.0,
-                    maxWidth: 60.0,
-                    maxHeight: (constraints?.maxHeight ?? 40).scaleSize),
-                suffixIconConstraints: BoxConstraints(
-                    minWidth: 10.0,
-                    maxWidth: 60.0,
-                    maxHeight: (constraints?.maxHeight ?? 40).scaleSize),
-                iconColor: white,
-                focusColor: white,
-                hintStyle: hintStyle,
-                suffixIcon: suffixIcon,
-              ),
-              maxLines: 1,
-              cursorHeight: ((constraints?.maxHeight ?? 35) - 15).scaleSize,
-              keyboardType: inputType ?? TextInputType.text,
-              onSubmitted: onSubmitted,
-              onChanged: onChanged,
-              autofocus: autoFocus ?? false,
-              obscureText: obscureText ?? false,
-              focusNode: focusNode,
-              controller: inputController,
-              enabled: enableInput,
-              style: textStyle)),
+          textCapitalization: TextCapitalization.words,
+          maxLines: maxLines,
+          cursorColor: cursorColor,
+          cursorErrorColor: warning,
+          keyboardType: inputType ?? TextInputType.text,
+          onSubmitted: onSubmitted,
+          onChanged: onChanged,
+          autofocus: autoFocus ?? false,
+          obscureText: obscureText ?? false,
+          focusNode: focusNode,
+          controller: inputController,
+          enabled: enableInput,
+          style: textStyle ?? defaultTextStyle),
     );
   }
 }